+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/__init__.py
@@ -75,6 +75,9 @@ from .country_specific.nigeria.ng_vehicle_registration_recognizer import (
     NgVehicleRegistrationRecognizer,
 )
 
+# Philippines recognizers
+from .country_specific.philippines.ph_tin_recognizer import PhTinRecognizer
+
 # Poland recognizers
 from .country_specific.poland.pl_pesel_recognizer import PlPeselRecognizer
 
@@ -212,6 +215,7 @@ __all__ = [
     "InPanRecognizer",
     "GLiNERRecognizer",
     "HuggingFaceNerRecognizer",
+    "PhTinRecognizer",
     "PlPeselRecognizer",
     "AzureAILanguageRecognizer",
     "InAadhaarRecognizer",